feat(bootstrap): collapse bootstrap fork to single orchestrator + remove config profiles (refs #3690)
- delete old bootstrap.js; rename bootstrap-new.js -> bootstrap.js with all
  [temp] development logging stripped and log prefixes normalized
- soften the preflight project-scope check: unreadable token scopes
  (fine-grained PATs, gh builds without a 'Token scopes:' line) now warn
  and pass instead of failing closed; a readable scopes line lacking
  'project' still fails with the refresh remedy
- remove the config-profile surface: .agents/config-profiles/,
  lib/config/profiles.js, buildProfileAgentrcBody, the profile seeding
  branch in ensureAgentrc, the profile explain source-kind, the --profile
  flag in mandrel explain, and the 'profile' KNOWN_FLAGS entry
- delete phased-approval.js (+ test) — only the old orchestrator used it;
  hitl-confirm.js / summary.js / manifest.js stay wired into
  agents-bootstrap-github.js
- repoint bootstrap entry tests at the renamed module; drop tests for
  removed surface (renderSummary/SECTIONS, profiles, profile selection)
- docs: project-board.md now describes the 4-view REST reality;
  agents-update.md drops the stale Step 7.5 reference
