--- .agents/config-profiles/audit-only.json
@@ -1,16 +1,0 @@
-{
-  "$schema": "../schemas/agentrc.schema.json",
-  "project": {
-    "baseBranch": "main",
-    "paths": {
-      "agentRoot": ".agents",
-      "docsRoot": "docs",
-      "tempRoot": "temp"
-    }
-  },
-  "delivery": {
-    "codeReview": {
-      "providers": [{ "name": "security-review", "optional": true }]
-    }
-  }
-}
