--- .agents/config-profiles/qa-only.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "../schemas/agentrc.schema.json",
-  "project": {
-    "baseBranch": "main",
-    "paths": {
-      "agentRoot": ".agents",
-      "docsRoot": "docs",
-      "tempRoot": "temp"
-    }
-  },
-  "qa": {
-    "featureRoot": "tests/features",
-    "signInSeam": {
-      "urlTemplate": "/dev/login?persona={persona}"
-    },
-    "personas": ["admin", "member"]
-  }
-}
