--- .agents/config-profiles/solo-local.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "../schemas/agentrc.schema.json",
-  "project": {
-    "baseBranch": "main",
-    "paths": {
-      "agentRoot": ".agents",
-      "docsRoot": "docs",
-      "tempRoot": "temp"
-    }
-  }
-}
