--- .agents/config-profiles/team-github.json
@@ -1,28 +1,0 @@
-{
-  "$schema": "../schemas/agentrc.schema.json",
-  "project": {
-    "baseBranch": "main",
-    "paths": {
-      "agentRoot": ".agents",
-      "docsRoot": "docs",
-      "tempRoot": "temp"
-    }
-  },
-  "github": {
-    "owner": "[OWNER]",
-    "repo": "[REPO]",
-    "operatorHandle": "@[USERNAME]",
-    "branchProtection": {
-      "enforce": true
-    },
-    "notifications": {
-      "mentionOperator": true,
-      "commentEvents": ["state-transition", "story-merged"]
-    }
-  },
-  "delivery": {
-    "ci": {
-      "skipForStoryPushes": true
-    }
-  }
-}
