+++ .agents/scripts/bootstrap.js
@@ -2,11 +2,21 @@
 /**
  * bootstrap.js — single-command consumer setup for Mandrel.
  *
- * Replaces the legacy `/agents-bootstrap-project` + `/agents-bootstrap-github`
- * slash commands with one deterministic Node script (Story #2074, hard
- * cutover). Walks the operator through every required value (auto-detecting
- * what it can), applies the project-side file mutations, then runs the
- * existing GitHub-side bootstrap.
+ * The sole bootstrap orchestrator (Story #3690 collapsed the temporary
+ * bootstrap-new.js fork into this file). Key behaviours:
+ *   - No config profiles — `.agentrc.json` always seeds from the bundled
+ *     `.agents/starter-agentrc.json` starter reference.
+ *   - Runs even when the directory is NOT a git repo yet (preflight detects
+ *     git state instead of failing on it).
+ *   - Adds a Projects V2 permission check to preflight (warns rather than
+ *     failing when classic token scopes cannot be read, e.g. fine-grained
+ *     PATs).
+ *   - Uses a plain summary + confirm loop (interactive runs can go back and
+ *     re-answer) instead of a phased-approval manifest.
+ *   - Provisions the missing pieces of a cold start: initializes the local
+ *     git repo (with a first commit) when absent, creates the GitHub repo
+ *     (linking + pushing the local tree), and creates the Projects V2 board
+ *     from a typed name — capturing its number for the rest of the run.
  *
  * Usage:
  *   node .agents/scripts/bootstrap.js [flags]
@@ -14,21 +24,27 @@
  * Flags:
  *   --owner <name>            GitHub owner (default: parsed from origin remote)
  *   --repo <name>             GitHub repo  (default: parsed from origin remote)
- *   --operator-handle <name>  GitHub handle for `github.operatorHandle`
- *   --base-branch <name>      Base branch (default: origin/HEAD or `main`)
- *   --project-number <n>      Projects V2 number (optional)
- *   --profile <name>          Config profile to seed .agentrc.json from
- *                             (solo-local|team-github|qa-only|audit-only);
- *                             blank uses the full starter reference
- *   --assume-yes              Accept every default; without it a non-TTY run declines all prompts
+ *   --visibility <v>          Visibility for a newly created repo:
+ *                             private | public | internal (default: private)
+ *   --operator-handle <name>  GitHub handle for github.operatorHandle
+ *   --base-branch <name>      Base branch (default: origin/HEAD or 'main')
+ *   --project-number <n>      Projects V2 number/name (optional)
+ *   --assume-yes              Accept every default; required for non-TTY runs
  *   --skip-github             Skip the GitHub-side bootstrap entirely
- *   --skip-quality            Skip the quality-gates bootstrap (Step 7.5)
- *   --dry-run                 Print the resolved plan without mutating
+ *   --skip-quality            Skip the quality-gates bootstrap
+ *   --dry-run                 Collect info and print the plan; change nothing
+ *   --reap-conflicting-workflows  Delete Projects V2 built-in workflows that
+ *                             race against the orchestrator (destructive)
  *   --help                    Print this help
  */
 
+import { spawnSync } from 'node:child_process';
+import fs from 'node:fs';
 import path from 'node:path';
+import readline from 'node:readline/promises';
 import { fileURLToPath } from 'node:url';
+
+// Reused bootstrap library helpers (unchanged).
 import { listProjects, listRepos } from './lib/bootstrap/gh-list.js';
 import {
   buildLedgerRecord,
@@ -38,7 +54,6 @@ import {
   buildMutationManifest,
   PHASE_GROUPS,
 } from './lib/bootstrap/manifest.js';
-import { runPhasedApproval } from './lib/bootstrap/phased-approval.js';
 import { runPreflight } from './lib/bootstrap/preflight.js';
 import { applyProjectBootstrap } from './lib/bootstrap/project-bootstrap.js';
 import {
@@ -47,7 +62,7 @@ import {
   parseFlags,
 } from './lib/bootstrap/prompt.js';
 import { runAsCli } from './lib/cli-utils.js';
-import { listProfiles, PROFILE_NAMES } from './lib/config/profiles.js';
+import { exec, GhNotFoundError } from './lib/gh-exec.js';
 import { Logger } from './lib/Logger.js';
 
 const HELP = `bootstrap.js — single-command consumer setup for Mandrel.
@@ -57,119 +72,404 @@ Usage: node .agents/scripts/bootstrap.js [flags]
 Flags:
   --owner <name>            GitHub owner (default: parsed from origin remote)
   --repo <name>             GitHub repo  (default: parsed from origin remote)
+  --visibility <v>          Visibility for a newly created repo:
+                            private | public | internal (default: private)
   --operator-handle <name>  GitHub handle for github.operatorHandle
   --base-branch <name>      Base branch (default: origin/HEAD or 'main')
-  --project-number <n>      Projects V2 number (optional)
-  --profile <name>          Config profile to seed .agentrc.json from
-                            (solo-local|team-github|qa-only|audit-only);
-                            blank uses the full starter reference
-  --assume-yes              Accept every default; without it a non-TTY run declines all prompts
+  --project-number <n>      Projects V2 number/name (optional)
+  --assume-yes              Accept every default; required for non-TTY runs
   --skip-github             Skip the GitHub-side bootstrap entirely
-  --skip-quality            Skip the quality-gates bootstrap (Step 7.5)
-  --dry-run                 Print the resolved plan without mutating
+  --skip-quality            Skip the quality-gates bootstrap
+  --dry-run                 Collect info and print the plan; change nothing
+  --reap-conflicting-workflows  Delete Projects V2 built-in workflows that
+                            race against the orchestrator (destructive)
   --help                    Print this help
 `;
 
-/**
- * Resolve the GitHub owner using the same flag → env → inferred-default
- * precedence the `owner` question itself walks (`resolveFromFlag` →
- * `resolveFromEnv` → `resolveFromSilent`). The repo / project pickers need
- * the owner up front to scope their `gh` queries, but the picker `list`
- * closures run before `collectAnswers` has recorded the owner answer, so we
- * recompute it here from the same sources rather than reading a not-yet-set
- * answer.
- *
- * Returns `null` when no owner can be determined — both pickers degrade to
- * an empty list (and therefore skip) in that case.
- *
- * @param {{ owner?: string|null }} defaults
- * @param {Record<string, string|boolean>} flags
- * @param {NodeJS.ProcessEnv} env
- * @returns {string|null}
- */
+// ---------------------------------------------------------------------------
+// Small helpers
+// ---------------------------------------------------------------------------
+
+/** Strip an `owner/` prefix off a repo slug, leaving the bare repo name. */
+function bareRepoName(slug) {
+  const slash = slug.indexOf('/');
+  return slash === -1 ? slug : slug.slice(slash + 1);
+}
+
+/** Run a list-producing fn, returning [] on any throw. */
+function safeList(fn) {
+  try {
+    return fn() ?? [];
+  } catch {
+    return [];
+  }
+}
+
+/** Resolve the GitHub owner for the pickers: flag → env → inferred default. */
 export function resolveOwnerForPicker(defaults, flags, env = process.env) {
-  const fromFlag = flags?.owner;
-  if (typeof fromFlag === 'string' && fromFlag.length > 0) return fromFlag;
-  const fromEnv = env?.GH_OWNER;
-  if (typeof fromEnv === 'string' && fromEnv.length > 0) return fromEnv;
-  const fromDefault = defaults?.owner;
-  if (typeof fromDefault === 'string' && fromDefault.length > 0) {
-    return fromDefault;
+  if (typeof flags?.owner === 'string' && flags.owner.length > 0) {
+    return flags.owner;
+  }
+  if (typeof env?.GH_OWNER === 'string' && env.GH_OWNER.length > 0) {
+    return env.GH_OWNER;
+  }
+  if (typeof defaults?.owner === 'string' && defaults.owner.length > 0) {
+    return defaults.owner;
   }
   return null;
 }
 
-/**
- * Build the question list for the interactive bootstrap prompts.
- *
- * The `repo` and `projectNumber` questions carry an optional
- * `picker: { list }` field that `resolveFromPicker` consumes when running
- * interactively. The repo picker lists the owner's repositories (bare repo
- * names, derived from each `owner/name` slug returned by `listRepos`); the
- * project picker lists the owner's Projects V2 titles via `listProjects`.
- *
- * Both pickers are interactive-only by virtue of `resolveFromPicker`'s
- * `ctx.interactive` gate and its position in the `RESOLVERS` chain (after
- * flag / env / silent resolvers): when `--owner`/`--repo` or
- * `GH_OWNER`/`GH_REPO` are supplied, or `--assume-yes` is set, the earlier
- * resolvers win and the picker never runs. The pickers also self-skip when
- * the owner cannot be resolved (empty list) so a missing owner never blocks.
- *
- * @param {{ owner?: string|null, repo?: string|null,
- *           baseBranch?: string|null, operatorHandle?: string|null }} defaults
- * @param {object} [pickerCtx]
- * @param {Record<string, string|boolean>} [pickerCtx.flags]
- * @param {NodeJS.ProcessEnv} [pickerCtx.env]
- * @param {{ listRepos?: typeof listRepos, listProjects?: typeof listProjects }}
- *   [pickerCtx.providers] Injectable `gh-list` providers (tests pass stubs).
- */
-export function buildQuestions(defaults, pickerCtx = {}) {
-  const flags = pickerCtx.flags ?? {};
-  const env = pickerCtx.env ?? process.env;
-  const providers = pickerCtx.providers ?? {};
-  const repoLister = providers.listRepos ?? listRepos;
-  const projectLister = providers.listProjects ?? listProjects;
-  const owner = resolveOwnerForPicker(defaults, flags, env);
+/** Ask a yes/no question. Non-interactive runs auto-accept (return true). */
+async function confirmYesNo(message, interactive) {
+  if (!interactive) return true;
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+  try {
+    const raw = (await rl.question(`${message} [Y/n]: `)).trim().toLowerCase();
+    return raw === '' || raw === 'y' || raw === 'yes';
+  } finally {
+    rl.close();
+  }
+}
 
+/**
+ * Run a git command in `cwd`. Returns the normalized
+ * `{ ok, status, stdout, stderr, error }` shape (mirroring the bootstrap
+ * preflight/gh-list runners) so callers branch on `ok` without juggling
+ * spawnSync internals.
+ */
+function runGit(args, cwd) {
+  const result = spawnSync('git', args, { cwd, encoding: 'utf8' });
+  return {
+    ok: !result.error && result.status === 0,
+    status: result.status,
+    stdout: typeof result.stdout === 'string' ? result.stdout.trim() : '',
+    stderr: typeof result.stderr === 'string' ? result.stderr.trim() : '',
+    error: result.error,
+  };
+}
+
+/**
+ * Surface a `gh`/exec failure with the same detail the GitHub bootstrap
+ * step prints — message plus the real gh stderr/stdout/args carried on a
+ * `GhExecError` — so a bare "gh exited with code 1" is actually diagnosable.
+ */
+function logGhError(label, err) {
+  Logger.error(`[bootstrap] ${label} failed: ${err.message}`);
+  if (err.stderr)
+    Logger.error(`[bootstrap]   gh stderr: ${String(err.stderr).trim()}`);
+  if (err.stdout)
+    Logger.error(`[bootstrap]   gh stdout: ${String(err.stdout).trim()}`);
+  if (Array.isArray(err.args)) {
+    Logger.error(`[bootstrap]   gh args: ${err.args.join(' ')}`);
+  }
+}
+
+/**
+ * Per-command git identity args. The first commit fails when neither a repo-
+ * nor global-level `user.name`/`user.email` is configured, which is common on
+ * a freshly provisioned machine. When either is missing we supply a
+ * non-persistent identity via `-c` (derived from the operator handle) so the
+ * commit succeeds without mutating the operator's git config.
+ */
+function gitIdentityArgs(cwd, answers) {
+  const haveName = runGit(['config', 'user.name'], cwd).ok;
+  const haveEmail = runGit(['config', 'user.email'], cwd).ok;
+  if (haveName && haveEmail) return [];
+  const handle = answers.operatorHandle || answers.owner || 'mandrel';
+  return [
+    '-c',
+    `user.name=${handle}`,
+    '-c',
+    `user.email=${handle}@users.noreply.github.com`,
+  ];
+}
+
+/**
+ * Initialize the local git repo when one is not already present, and ensure
+ * at least one commit exists so `gh repo create --source=. --push` has
+ * something to push. Idempotent: a repo that already resolves `HEAD` is left
+ * untouched. Returns `{ ok, initialized, committed }` (or `{ ok:false, error }`
+ * on failure).
+ */
+function ensureGitInitialized(state) {
+  const cwd = state.projectRoot;
+  const branch = state.answers.baseBranch || 'main';
+  let initialized = false;
+  if (!state.gitInitialized) {
+    // `git init -b <branch>` (git ≥ 2.28) sets the initial branch directly;
+    // fall back to a plain init + symbolic-ref for older git.
+    let init = runGit(['init', '-b', branch], cwd);
+    if (!init.ok) {
+      init = runGit(['init'], cwd);
+      if (!init.ok)
+        return { ok: false, error: init.stderr || 'git init failed' };
+      runGit(['symbolic-ref', 'HEAD', `refs/heads/${branch}`], cwd);
+    }
+    initialized = true;
+    state.gitInitialized = true;
+    Logger.info(
+      `[bootstrap] Initialized git repo (branch ${branch}) at ${cwd}.`,
+    );
+  }
+
+  // A push needs a commit; create one only when HEAD does not resolve yet.
+  let committed = false;
+  if (!runGit(['rev-parse', '--verify', 'HEAD'], cwd).ok) {
+    runGit(['add', '-A'], cwd);
+    const commit = runGit(
+      [
+        ...gitIdentityArgs(cwd, state.answers),
+        'commit',
+        '--allow-empty',
+        '-m',
+        'Initial commit',
+      ],
+      cwd,
+    );
+    if (!commit.ok) {
+      return { ok: false, error: commit.stderr || 'git commit failed' };
+    }
+    committed = true;
+    Logger.info('[bootstrap] Created initial commit.');
+  }
+  return { ok: true, initialized, committed };
+}
+
+/**
+ * Wire the local `origin` remote to owner/repo when it is missing, so the
+ * GitHub bootstrap — which infers the target repo from the local remote —
+ * can run. This is the companion to `createGithubRepo`: that path wires
+ * `origin` itself via `--remote origin`, but a repo that already exists (or
+ * a re-run after a partial failure) leaves the local folder unlinked. Only
+ * acts when the repo actually exists on GitHub; pushes the base branch to
+ * set upstream, downgrading a rejected push to a warning since the bootstrap
+ * only needs the remote to resolve (content sync is the operator's to settle).
+ */
+async function ensureGitRemote(state) {
+  const cwd = state.projectRoot;
+  const { owner, repo } = state.answers;
+  const branch = state.answers.baseBranch || 'main';
+  if (runGit(['remote', 'get-url', 'origin'], cwd).ok) return;
+  if (!(await repoExists(owner, repo))) {
+    Logger.warn(
+      `[bootstrap] No 'origin' remote and ${owner}/${repo} does not exist on GitHub — skipping remote wiring.`,
+    );
+    return;
+  }
+  const url = `https://github.com/${owner}/${repo}.git`;
+  const add = runGit(['remote', 'add', 'origin', url], cwd);
+  if (!add.ok) {
+    Logger.warn(`[bootstrap] Could not add 'origin' remote: ${add.stderr}`);
+    return;
+  }
+  Logger.info(`[bootstrap] Wired 'origin' → ${url}.`);
+  const push = runGit(['push', '-u', 'origin', branch], cwd);
+  if (!push.ok) {
+    Logger.warn(
+      `[bootstrap] 'origin' is set but push of '${branch}' failed (resolve manually, e.g. \`git pull --rebase origin ${branch}\`): ${push.stderr}`,
+    );
+  }
+}
+
+/**
+ * Authoritatively check whether `owner/repo` exists, via `gh repo view`.
+ * Returns false on a not-found (so the repo can be created), true when it
+ * resolves, and true on any other error (auth/network/etc.) so a transient
+ * failure never triggers a spurious create attempt. Used instead of an
+ * `is it in the repo-list?` heuristic, which mis-fires for a brand-new
+ * account whose `gh repo list` is empty.
+ */
+async function repoExists(owner, repo) {
+  try {
+    await exec({
+      args: ['repo', 'view', `${owner}/${repo}`, '--json', 'name'],
+    });
+    return true;
+  } catch (err) {
+    if (err instanceof GhNotFoundError) return false;
+    return true;
+  }
+}
+
+/**
+ * Link the repo to the Projects V2 board (`gh project link`) so issues and
+ * PRs from the repo surface on the project. Runs whenever both a numeric
+ * project number and a repo are resolved — a freshly created project or an
+ * existing one picked from the list. Non-fatal and re-run-safe: an
+ * already-linked repo or a transient hiccup is downgraded to a warning so it
+ * never fails the bootstrap.
+ */
+async function ensureProjectLinked(state) {
+  const { owner, repo } = state.answers;
+  const pn = String(state.answers.projectNumber ?? '');
+  if (!/^\d+$/.test(pn) || !repo) return;
+  try {
+    await exec({
+      args: ['project', 'link', pn, '--owner', owner, '--repo', repo],
+    });
+    Logger.info(
+      `[bootstrap] Linked repo ${owner}/${repo} to Project V2 #${pn}.`,
+    );
+  } catch (err) {
+    Logger.warn(
+      `[bootstrap] Could not link repo ${owner}/${repo} to Project V2 #${pn} (continuing): ${err.message}`,
+    );
+  }
+}
+
+/** Visibilities `gh repo create` accepts; each maps to a `--<v>` flag. */
+export const REPO_VISIBILITIES = Object.freeze([
+  'private',
+  'public',
+  'internal',
+]);
+
+/**
+ * Resolve the new-repo visibility from `--visibility` (default `private`).
+ * Case-insensitive. Returns `null` for an unrecognized value so the caller
+ * can reject it with a clear message instead of silently defaulting.
+ */
+export function resolveRepoVisibility(flags = {}) {
+  const raw = flags.visibility;
+  if (typeof raw !== 'string' || raw.length === 0) return 'private';
+  const value = raw.trim().toLowerCase();
+  return REPO_VISIBILITIES.includes(value) ? value : null;
+}
+
+/**
+ * Create the GitHub repo from the resolved owner/repo. `--source` links the
+ * existing local repo, `--remote origin` wires the remote, and `--push`
+ * uploads the current branch — so the local tree and the new remote stay in
+ * lockstep and Step 1's auto-detection works on a re-run. Visibility comes
+ * from `--visibility` (default private). Throws GhExecError on failure
+ * (surfaced by the caller).
+ */
+async function createGithubRepo(state) {
+  const { owner, repo } = state.answers;
+  const slug = `${owner}/${repo}`;
+  const visibility = resolveRepoVisibility(state.flags);
+  await exec({
+    args: [
+      'repo',
+      'create',
+      slug,
+      `--${visibility}`,
+      '--source',
+      state.projectRoot,
+      '--remote',
+      'origin',
+      '--push',
+    ],
+  });
+  Logger.info(
+    `[bootstrap] Created GitHub repo ${slug} (${visibility}) and pushed.`,
+  );
+}
+
+/**
+ * Create a Projects V2 board from the typed name and rewrite
+ * `state.answers.projectNumber` to the assigned numeric id so the downstream
+ * persist + GitHub bootstrap steps treat it as an existing project (and never
+ * create a duplicate). Throws on failure or when gh returns no number.
+ */
+async function createGithubProject(state) {
+  const { owner } = state.answers;
+  const title = String(state.answers.projectNumber);
+  // `gh project create` uses `--format json` (not `--json`), so exec returns
+  // the raw `{ stdout }` envelope — parse the number ourselves.
+  const res = await exec({
+    args: [
+      'project',
+      'create',
+      '--owner',
+      owner,
+      '--title',
+      title,
+      '--format',
+      'json',
+    ],
+  });
+  let number = null;
+  try {
+    number = JSON.parse(res.stdout)?.number ?? null;
+  } catch {
+    /* fall through to the guard below */
+  }
+  if (!Number.isInteger(number)) {
+    throw new Error(
+      `gh project create returned no numeric project number (stdout: ${res.stdout?.trim() ?? ''})`,
+    );
+  }
+  state.answers.projectNumber = String(number);
+  Logger.info(`[bootstrap] Created GitHub Project V2 "${title}" (#${number}).`);
+  return number;
+}
+
+// ---------------------------------------------------------------------------
+// Question list (Step 3).
+// ---------------------------------------------------------------------------
+
+/**
+ * Build the Step 3 question list. `silentAccept` keys (owner/repo/baseBranch/
+ * operatorHandle) are git-inferred and accepted without prompting unless an
+ * override is supplied. The `operatorHandle` and `projectNumber` defaults
+ * track the repo owner / repo name respectively (see post-processing in
+ * `collectAndConfirm`).
+ */
+export function buildQuestions(defaults, flags, env = process.env, lists = {}) {
+  const owner = resolveOwnerForPicker(defaults, flags, env);
+  // Pre-fetched lists (shared with the summary display) are a fast path used
+  // only when populated. They're empty when the owner is unknown up front
+  // (a folder with no git remote), so the pickers fall back to a live fetch
+  // keyed off the owner the operator just typed (`answers.owner`).
+  const reposList = lists.reposList;
+  const projectsList = lists.projectsList;
+  const pickerOwner = (answers) => answers?.owner || owner;
   return [
     {
       key: 'owner',
       flag: 'owner',
       env: 'GH_OWNER',
-      message: 'GitHub owner',
+      message: 'Github repo owner',
       default: defaults.owner,
       required: true,
       validate: (v) =>
         /^[A-Za-z0-9][A-Za-z0-9-]*$/.test(v) ? null : 'Invalid GitHub owner',
     },
     {
-      key: 'repo',
-      flag: 'repo',
-      env: 'GH_REPO',
-      message: 'GitHub repo',
-      default: defaults.repo,
-      required: true,
-      picker: {
-        list: () =>
-          repoLister({ owner }).map((slug) => {
-            const slash = slug.indexOf('/');
-            return slash === -1 ? slug : slug.slice(slash + 1);
-          }),
-      },
-      validate: (v) =>
-        /^[A-Za-z0-9._-]+$/.test(v) ? null : 'Invalid GitHub repo name',
-    },
-    {
       key: 'operatorHandle',
       flag: 'operator-handle',
       env: 'GH_OPERATOR_HANDLE',
-      message: 'Operator GitHub handle (without @)',
-      default: defaults.operatorHandle,
+      message: 'Github username/handle (without the @)',
+      // Default tracks the repo owner; resolved post-collect if left blank.
+      default: defaults.owner,
       required: false,
       validate: (v) =>
         v.length === 0 || /^[A-Za-z0-9-]+$/.test(v)
           ? null
           : 'Invalid GitHub handle',
+    },
+    {
+      key: 'repo',
+      flag: 'repo',
+      env: 'GH_REPO',
+      message:
+        'Github repo name - Select from the list or enter a new name to create one',
+      default: defaults.repo,
+      required: true,
+      picker: {
+        list: (answers) => {
+          if (Array.isArray(reposList) && reposList.length > 0)
+            return reposList;
+          const o = pickerOwner(answers);
+          return o ? listRepos({ owner: o }).map(bareRepoName) : [];
+        },
+      },
+      validate: (v) =>
+        /^[A-Za-z0-9._-]+$/.test(v) ? null : 'Invalid GitHub repo name',
     },
     {
       key: 'baseBranch',
@@ -184,37 +484,26 @@ export function buildQuestions(defaults, pickerCtx = {}) {
       key: 'projectNumber',
       flag: 'project-number',
       env: 'GH_PROJECT_NUMBER',
-      message: 'Projects V2 number (blank to skip)',
-      default: null,
+      message:
+        'Github Project V2 name - Select from the list or enter a new name to create one',
+      // Default tracks the repo name; resolved post-collect if left blank.
+      default: defaults.repo,
       required: false,
       picker: {
-        list: () => projectLister({ owner }),
+        list: (answers) => {
+          if (Array.isArray(projectsList) && projectsList.length > 0) {
+            return projectsList;
+          }
+          const o = pickerOwner(answers);
+          return o ? listProjects({ owner: o }) : [];
+        },
       },
+      // Accept blank (skip), an existing project number, or a new project
+      // name (letters/digits/space/._-).
       validate: (v) =>
-        v.length === 0 || /^\d+$/.test(v)
+        v.length === 0 || /^\d+$/.test(v) || /^[A-Za-z0-9 ._-]+$/.test(v)
           ? null
-          : 'Must be an integer or blank',
-    },
-    {
-      // Story #3527 — pick a named config profile to seed `.agentrc.json`
-      // from that profile's posture-scoped delta rather than the full
-      // bundled starter reference. Blank keeps the historical default (the
-      // full starter). The picker lists the canonical profile names; the
-      // validator rejects any non-empty value that is not a known profile so
-      // a typo never reaches `ensureAgentrc`.
-      key: 'profile',
-      flag: 'profile',
-      env: 'MANDREL_PROFILE',
-      message: 'Config profile (blank for the full starter reference)',
-      default: null,
-      required: false,
-      picker: {
-        list: () => listProfiles(),
-      },
-      validate: (v) =>
-        v.length === 0 || PROFILE_NAMES.includes(v)
-          ? null
-          : `Unknown profile. Known profiles: ${PROFILE_NAMES.join(', ')}.`,
+          : 'Invalid project name',
     },
   ];
 }
@@ -225,14 +514,12 @@ const INFERRED_KEYS = Object.freeze([
   'baseBranch',
   'operatorHandle',
 ]);
-
 const FLAG_BY_KEY = Object.freeze({
   owner: 'owner',
   repo: 'repo',
   baseBranch: 'base-branch',
   operatorHandle: 'operator-handle',
 });
-
 const ENV_BY_KEY = Object.freeze({
   owner: 'GH_OWNER',
   repo: 'GH_REPO',
@@ -240,19 +527,7 @@ const ENV_BY_KEY = Object.freeze({
   operatorHandle: 'GH_OPERATOR_HANDLE',
 });
 
-/**
- * Compute the set of keys whose inferred default should be accepted
- * without prompting. A key qualifies when:
- *   - `inferDefaults` produced a non-empty value for it; AND
- *   - no CLI flag was supplied for it; AND
- *   - no environment variable override was supplied for it.
- *
- * Flag / env overrides win, but they take their normal path through
- * `collectAnswers` rather than this silent-accept set, so the operator
- * still sees them logged in the per-field resolution.
- *
- * Exported for unit testing.
- */
+/** Keys whose git-inferred default is accepted without prompting. */
 export function resolveSilentAccept(defaults, flags, env = process.env) {
   const out = [];
   for (const key of INFERRED_KEYS) {
@@ -266,160 +541,8 @@ export function resolveSilentAccept(defaults, flags, env = process.env) {
 }
 
 // ---------------------------------------------------------------------------
-// Summary sections (Story #2459 / Task #2472)
-//
-// `printSummary` used to be a 47-line conditional ladder with three `if/else`
-// pairs interleaved with linear `Logger.info` calls. Each line is now a
-// declarative `{ label, render }` entry — the renderer returns either:
-//
-//   - `string`     — exact `Logger.info` argument (callers prepend their
-//                    own indentation, banner `\n`, etc.).
-//   - `string[]`   — one `Logger.info` call per entry.
-//   - `null`       — suppress this section entirely.
-//
-// `printSummary` becomes a single one-branch loop. The strings are
-// pre-padded inside each renderer rather than computed via a generic
-// pad-to-width because the original implementation hand-tuned column
-// widths per logical section (the github lines use a different column
-// width than the project lines). Snapshot tests pin the exact output.
-//
-// Exported for the snapshot test in `tests/scripts/bootstrap-print-summary.test.js`.
+// GitHub-side bootstrap (Step 6) — same wiring as bootstrap.js.
 // ---------------------------------------------------------------------------
-
-// Every renderer returns `string[]` (or `null` to suppress the section)
-// so the flatten step stays a single-branch loop.
-
-export const SECTIONS = Object.freeze([
-  {
-    name: 'banner',
-    render: () => ['\n=== Bootstrap Summary ==='],
-  },
-  {
-    name: 'package.json',
-    render: (r) => [
-      `  package.json           created=${r.pkg.created} sync:commands=${r.pkg.scriptsSyncCommands} prepare=${r.pkg.scriptsPrepare} bootstrap=${r.pkg.scriptsBootstrap}`,
-    ],
-  },
-  {
-    name: 'install',
-    render: (r) => [
-      `  install                ${r.install.ran ? `ran via ${r.install.manager}` : `skipped (${r.install.reason})`}`,
-    ],
-  },
-  {
-    name: 'agentrc',
-    render: (r) => [`  .agentrc.json          ${r.agentrc.action}`],
-  },
-  {
-    name: 'claudeSettings',
-    render: (r) => [`  .claude/settings.json  ${r.claudeSettings.action}`],
-  },
-  {
-    name: 'systemPromptWiring',
-    render: (r) => [`  CLAUDE.md              ${r.systemPromptWiring.action}`],
-  },
-  {
-    name: 'gitignore.commands',
-    render: (r) => [`  .gitignore commands    ${r.gitignore.commands}`],
-  },
-  {
-    name: 'gitignore.mcp',
-    render: (r) => [`  .gitignore mcp         ${r.gitignore.mcp}`],
-  },
-  {
-    name: 'parity',
-    render: (r) => [`  parity                 ${r.parity.ok ? 'OK' : 'FAIL'}`],
-  },
-  {
-    name: 'quality',
-    render: (r) =>
-      r.quality.skipped
-        ? ['  quality                skipped']
-        : [
-            `  quality.helper         ${r.quality.helper.action}`,
-            `  quality.hook           ${r.quality.hook.action}`,
-            `  quality.scripts        ${r.quality.scripts.action}`,
-            `  quality.config         ${r.quality.config.action}`,
-          ],
-  },
-  {
-    name: 'winPerf',
-    render: (r) =>
-      r.winPerf.skipped
-        ? [`  windows-git-perf       skipped (${r.winPerf.platform})`]
-        : [`  windows-git-perf       ${r.winPerf.ok ? 'OK' : 'warnings'}`],
-  },
-  {
-    name: 'github',
-    render: (r) => {
-      if (!r.github) return ['  github                 skipped'];
-      if (r.github.error) {
-        return [`  github                 error (${r.github.error})`];
-      }
-      if (r.github.skipped) {
-        return [
-          `  github                 skipped (${r.github.reason ?? 'skipped'})`,
-        ];
-      }
-      return [
-        `  github.labels          created=${r.github.labels.created.length} skipped=${r.github.labels.skipped.length}`,
-        `  github.project         ${r.github.project.projectNumber ?? 'skipped'}`,
-        `  github.branchProtection ${r.github.branchProtection.status ?? 'n/a'}`,
-        `  github.mergeMethods    ${r.github.mergeMethods.status ?? 'n/a'}`,
-      ];
-    },
-  },
-  {
-    name: 'ledger',
-    render: (r) => {
-      if (!r.ledger) return null;
-      return r.ledger.written
-        ? [
-            `  install-ledger         wrote ${r.ledger.entryCount} entr${r.ledger.entryCount === 1 ? 'y' : 'ies'} (groups: ${r.ledger.approvedGroups.join(', ')})`,
-          ]
-        : [`  install-ledger         not written (${r.ledger.reason})`];
-    },
-  },
-]);
-
-/**
- * Render the full summary as one newline-joined string. Pure helper so
- * snapshot tests can compare against fixture output without intercepting
- * `Logger.info` calls. Each section's renderer is responsible for its own
- * indentation and column widths — `renderSummary` only stitches results
- * together. Exported for tests.
- *
- * @param {object} report
- * @returns {string}
- */
-export function renderSummary(report) {
-  return flattenSummaryLines(report).join('\n');
-}
-
-/**
- * Flatten the section list to a sequence of strings (one per emitted
- * line). Each entry becomes one `Logger.info` call so the log adapter
- * sees the same line count it did before the refactor.
- *
- * Exported for the snapshot test (the assertion runs against this list
- * directly when the input is a fixture report).
- *
- * @param {object} report
- * @returns {string[]}
- */
-export function flattenSummaryLines(report) {
-  const out = [];
-  for (const section of SECTIONS) {
-    const lines = section.render(report);
-    if (lines == null) continue;
-    out.push(...lines);
-  }
-  return out;
-}
-
-function printSummary(report) {
-  for (const line of flattenSummaryLines(report)) Logger.info(line);
-}
 
 async function runGithubBootstrap(answers, opts) {
   const { runBootstrap, preflightGh, preflightRuntimeDeps } = await import(
@@ -430,6 +553,10 @@ async function runGithubBootstrap(answers, opts) {
   const { resolveConfig, validateOrchestrationConfig } = await import(
     './lib/config-resolver.js'
   );
+  // `resolveConfig` reads github.projectNumber from .agentrc.json, which the
+  // persistProjectNumber step writes BEFORE this runs — so the provider reuses
+  // the existing project instead of creating a new one (Bug: gave #8, created
+  // #12 because the number never reached the provider config).
   const config = resolveConfig();
   validateOrchestrationConfig(config);
   return runBootstrap(config, {
@@ -437,57 +564,43 @@ async function runGithubBootstrap(answers, opts) {
     github: config.github,
     assumeYes: opts.assumeYes,
     baseBranch: answers.baseBranch,
-    // Story #3526 — GitHub-admin mutations are explicit opt-in inside
-    // `runBootstrap`. This call site is reached only from
-    // `executeGithubBootstrap`, which already gates on the `github-admin`
-    // phase group being approved (and on `--skip-github` being absent), so
-    // arriving here IS the operator's consent. Pass it through so the
-    // boundary-level gate applies the approved mutations rather than
-    // short-circuiting to a no-op.
     githubAdminApproved: true,
+    // Opt-in: delete the Projects V2 built-in workflows that race against the
+    // orchestrator's ColumnSync (e.g. "Pull request merged"). Off by default.
+    reapConflictingWorkflows: Boolean(opts.reapConflictingWorkflows),
   });
 }
 
+/** True only when every github-admin sub-mutation that ran succeeded. */
+function githubSubMutationsSucceeded(gh) {
+  if (gh.branchProtection?.status === 'failed') return false;
+  if (gh.mergeMethods?.status === 'failed') return false;
+  return true;
+}
+
+/** Phase groups whose mutations actually landed, for the install ledger. */
+function resolveAppliedGroups(approvedGroups, report) {
+  const applied = new Set();
+  for (const group of approvedGroups ?? []) {
+    if (group === PHASE_GROUPS.GITHUB_ADMIN) {
+      const gh = report?.github;
+      if (gh && !gh.error && !gh.skipped && githubSubMutationsSucceeded(gh)) {
+        applied.add(group);
+      }
+      continue;
+    }
+    applied.add(group);
+  }
+  return applied;
+}
+
 // ---------------------------------------------------------------------------
-// Phase helpers (Story #2459 / Task #2471)
-//
-// `main()` was a 75-line procedural pipeline that mixed flag parsing,
-// non-TTY validation, defaults inference, answer collection, the project-
-// side mutation, and the GitHub-side bootstrap into a single function
-// with seven decision points. Each phase below is now a small helper that
-// returns either `{ ok: true, payload? }` (the pipeline advances) or
-// `{ ok: false, exit: <code> }` (the pipeline short-circuits with the
-// supplied exit code). `main` becomes a five-line pipeline driver.
-//
-// The helpers are exported so the sibling test file can exercise each
-// phase in isolation without spawning a child process.
+// Pipeline phases
 // ---------------------------------------------------------------------------
 
 /**
- * @typedef {object} PhaseHalt
- * @property {false} ok        — pipeline must stop here.
- * @property {number} exit     — process exit code to return.
- *
- * @typedef {object} PhaseAdvance
- * @property {true} ok         — pipeline advances to the next phase.
- * @property {object} [payload] — values the next phase needs.
- *
- * @typedef {PhaseHalt | PhaseAdvance} PhaseResult
- */
-
-/**
- * Phase 1 — Parse argv, short-circuit on `--help`, and enforce the
- * non-TTY contract (`--assume-yes` plus `--owner`/`--repo` or their env
- * equivalents).
- *
- * Exported for tests.
- *
- * @param {string[]} argv
- * @param {object} [opts]
- * @param {NodeJS.WritableStream} [opts.stdout]
- * @param {NodeJS.ProcessEnv} [opts.env]
- * @param {{ isTTY?: boolean }} [opts.stdin]
- * @returns {PhaseResult}
+ * Step 1 — Parse argv, handle `--help`, and enforce the non-TTY contract
+ * (`--owner`/`--repo` or env equivalents + `--assume-yes`).
  */
 export function parseAndValidate(argv, opts = {}) {
   const stdout = opts.stdout ?? process.stdout;
@@ -514,47 +627,19 @@ export function parseAndValidate(argv, opts = {}) {
       return { ok: false, exit: 1 };
     }
   }
+  if (resolveRepoVisibility(flags) === null) {
+    Logger.error(
+      `[bootstrap] invalid --visibility "${flags.visibility}". ` +
+        `Expected one of: ${REPO_VISIBILITIES.join(', ')}.`,
+    );
+    return { ok: false, exit: 1 };
+  }
   return { ok: true, payload: { flags, interactive, assumeYes } };
 }
 
 /**
- * Phase 1.5 — Unified prerequisite preflight (Story #3375). Runs as the
- * FIRST mutation-relevant gate: ahead of `prepareContext` and every
- * project-side / GitHub-side mutation. Aggregates the Node, git, and
- * (unless `--skip-github`) `gh` CLI + auth checks. On failure it prints
- * each failing check's remedy and halts the pipeline with exit code 1
- * before any mutation occurs.
- *
- * Exported for tests.
- *
- * @param {{ flags: Record<string, string|boolean> }} state
- * @param {{ run?: typeof runPreflight }} [opts]
- * @returns {Promise<PhaseResult>}
- */
-export async function runPreflightPhase(state, opts = {}) {
-  const run = opts.run ?? runPreflight;
-  const result = await run({ skipGithub: Boolean(state.flags['skip-github']) });
-  if (!result.ok) {
-    Logger.error(
-      '[bootstrap] Prerequisite preflight failed. Resolve the following before re-running:',
-    );
-    for (const check of result.checks) {
-      if (check.ok) continue;
-      Logger.error(`  - ${check.name}: ${check.remedy}`);
-    }
-    return { ok: false, exit: 1 };
-  }
-  return { ok: true, payload: { preflight: result } };
-}
-
-/**
- * Phase 2 — Resolve project paths, infer defaults from git, and compute
- * which inferred keys can be silently accepted. Pure I/O against the
- * local filesystem (no network calls).
- *
- * @param {{ flags: Record<string, string|boolean>, interactive: boolean }} state
- * @param {{ scriptUrl?: string, projectRoot?: string }} [opts]
- * @returns {PhaseAdvance}
+ * Step 1b — Resolve paths, infer defaults from git, and echo the detected
+ * values back to the operator (the "share found values" requirement).
  */
 export function prepareContext(state, opts = {}) {
   const scriptUrl = opts.scriptUrl ?? import.meta.url;
@@ -563,15 +648,13 @@ export function prepareContext(state, opts = {}) {
   const agentRoot = path.resolve(here, '..');
   const defaults = inferDefaults(projectRoot);
   const silentAccept = resolveSilentAccept(defaults, state.flags);
-  if (state.interactive && silentAccept.length > 0) {
-    const summary = silentAccept
-      .map((key) => `${key}=${defaults[key]}`)
-      .join(' ');
-    Logger.info(`[bootstrap] Auto-detected from local git: ${summary}`);
-    Logger.info(
-      '[bootstrap] Override any value with --owner / --repo / --base-branch / --operator-handle.',
-    );
-  }
+
+  Logger.info('[bootstrap] Detected from local git:');
+  Logger.info(`  owner          ${defaults.owner ?? '(none)'}`);
+  Logger.info(`  repo           ${defaults.repo ?? '(none)'}`);
+  Logger.info(`  base branch    ${defaults.baseBranch ?? '(none)'}`);
+  Logger.info(`  username       ${defaults.operatorHandle ?? '(none)'}`);
+
   return {
     ok: true,
     payload: { projectRoot, agentRoot, defaults, silentAccept },
@@ -579,232 +662,396 @@ export function prepareContext(state, opts = {}) {
 }
 
 /**
- * Phase 3 — Collect answers via the `RESOLVERS` chain and bail when any
- * required answer is missing. Also short-circuits the `--dry-run` plan
- * print so the operator can preview without mutating.
- *
- * @param {{
- *   flags: Record<string, string|boolean>,
- *   interactive: boolean,
- *   assumeYes: boolean,
- *   defaults: object,
- *   silentAccept: string[],
- * }} state
- * @returns {Promise<PhaseResult>}
+ * Step 2 — Preflight. Work-tree check is informational (does not fail the
+ * gate); adds the Projects V2 permission check. Prints a pass/fail line for
+ * every check.
  */
-export async function collectAndValidateAnswers(state) {
-  const { answers, missing } = await collectAnswers({
-    questions: buildQuestions(state.defaults, { flags: state.flags }),
-    flags: state.flags,
-    interactive: state.interactive,
-    assumeYes: state.assumeYes,
-    silentAccept: state.silentAccept,
+export async function runPreflightPhase(state, opts = {}) {
+  const run = opts.run ?? runPreflight;
+  const skipGithub = Boolean(state.flags['skip-github']);
+  const result = await run({
+    skipGithub,
+    requireWorkTree: false,
+    checkProjectScope: !skipGithub,
   });
-  if (missing.length > 0) {
-    Logger.error(`[bootstrap] missing required answers: ${missing.join(', ')}`);
+
+  for (const check of result.checks) {
+    if (check.ok) {
+      Logger.info(
+        `[bootstrap] ✓ ${check.name}${check.detail ? ` — ${check.detail}` : ''}`,
+      );
+    } else {
+      Logger.error(`[bootstrap] ✗ ${check.name}: ${check.remedy}`);
+    }
+  }
+
+  if (!result.ok) {
+    Logger.error(
+      '[bootstrap] Preflight failed. Resolve the issues above and re-run.',
+    );
     return { ok: false, exit: 1 };
   }
-  if (state.flags['dry-run']) {
-    // The dry-run plan is the FULL mutation manifest — the same single
-    // source the consent-first screen renders and execution honours
-    // (Story #3524). It includes the GitHub-side (`github-admin`) entries
-    // unless `--skip-github` was passed, so the operator previews every
-    // mutation, and it writes no files.
-    const manifestCtx = {
-      answers,
-      skipGithub: Boolean(state.flags['skip-github']),
-      skipQuality: Boolean(state.flags['skip-quality']),
-    };
-    Logger.info('[bootstrap] dry-run plan (no writes):');
+
+  Logger.info(
+    `[bootstrap] git initialized: ${result.gitInitialized ? 'yes' : 'no'}`,
+  );
+  return {
+    ok: true,
+    payload: { preflight: result, gitInitialized: result.gitInitialized },
+  };
+}
+
+/** Render the resolved answers as a human-readable summary block. */
+function renderAnswerSummary(
+  answers,
+  creation,
+  project,
+  gitInitialized,
+  visibility,
+) {
+  const newRepoNote = creation.newRepo
+    ? `  (NEW — will be created, ${visibility})`
+    : '';
+  const lines = [
+    '\n=== Review your answers ===',
+    `  Repo owner       ${answers.owner}`,
+    `  Username/handle  ${answers.operatorHandle || '(none)'}`,
+    `  Repo name        ${answers.repo}${newRepoNote}`,
+    `  Base branch      ${answers.baseBranch}`,
+    `  Project V2 name  ${project.name}${creation.newProject ? '  (NEW — will be created)' : ''}`,
+    `  Project V2 #     ${project.number}`,
+    `  Local git        ${gitInitialized ? 'initialized' : 'will be initialized'}`,
+  ];
+  return lines.join('\n');
+}
+
+/**
+ * Determine whether the answers ask for resources that do not exist yet.
+ * The repo is "new" when `gh repo view owner/repo` reports it does not
+ * exist — an authoritative per-repo probe rather than an "is it in the
+ * repo-list?" check, which mis-fired for a brand-new account whose
+ * `gh repo list` is empty (it then assumed the repo already existed and
+ * skipped creation). A non-numeric project answer (a typed name, not a
+ * picked number) is "new". When GitHub is skipped there is nothing to
+ * create, so detection is bypassed.
+ */
+async function detectCreation(answers, skipGithub) {
+  const creation = { newRepo: false, newProject: false };
+  if (skipGithub) return creation;
+  if (answers.repo && answers.owner) {
+    creation.newRepo = !(await repoExists(answers.owner, answers.repo));
+  }
+  const pn = answers.projectNumber;
+  if (typeof pn === 'string' && pn.length > 0 && !/^\d+$/.test(pn)) {
+    creation.newProject = true;
+  }
+  return creation;
+}
+
+/**
+ * Resolve the Projects V2 answer into a `{ name, number }` pair for the
+ * summary. The picker stores only the numeric value, so for an existing
+ * project (numeric answer) we look the name up in the owner's project list.
+ * A typed answer (non-numeric) is a new project name with no number yet.
+ */
+function resolveProjectDisplay(answers, skipGithub, projectsList) {
+  const pn = answers.projectNumber;
+  if (!pn) return { name: '(skip)', number: '(skip)' };
+  if (/^\d+$/.test(pn)) {
+    let name = '(unknown)';
+    if (!skipGithub) {
+      const projects =
+        projectsList ?? safeList(() => listProjects({ owner: answers.owner }));
+      const match = projects.find((p) => p.value === pn);
+      if (match) {
+        const m = /^(.*)\s+\(#\d+\)$/.exec(match.label);
+        name = m ? m[1] : match.label;
+      }
+    }
+    return { name, number: pn };
+  }
+  // Typed name → new project; number is assigned at creation time.
+  return { name: pn, number: '(new)' };
+}
+
+/**
+ * Steps 3 + 4 — Collect answers, show a summary, and confirm. Interactive
+ * runs that answer "no" loop back and re-ask. Non-interactive runs
+ * auto-accept. Then collect creation approval when a new repo/project was
+ * requested.
+ */
+export async function collectAndConfirm(state) {
+  const skipGithub = Boolean(state.flags['skip-github']);
+  const owner = resolveOwnerForPicker(state.defaults, state.flags);
+  // Fetch the owner's repos + projects ONCE and reuse for the pickers and
+  // the summary display — so the resolved project name never depends on a
+  // second (flaky) `gh` call. (Repo existence for the creation check is a
+  // separate, authoritative `gh repo view` probe in `detectCreation`.)
+  const reposList =
+    !skipGithub && owner
+      ? safeList(() => listRepos({ owner }).map(bareRepoName))
+      : [];
+  const projectsList =
+    !skipGithub && owner ? safeList(() => listProjects({ owner })) : [];
+
+  let silentAccept = state.silentAccept;
+  // Loop until the operator confirms the summary (or we auto-accept).
+  for (;;) {
+    const { answers, missing } = await collectAnswers({
+      questions: buildQuestions(state.defaults, state.flags, process.env, {
+        reposList,
+        projectsList,
+      }),
+      flags: state.flags,
+      interactive: state.interactive,
+      assumeYes: state.assumeYes,
+      silentAccept,
+    });
+    if (missing.length > 0) {
+      Logger.error(
+        `[bootstrap] missing required answers: ${missing.join(', ')}`,
+      );
+      return { ok: false, exit: 1 };
+    }
+    // Defaults that track another answer: handle ⇐ owner, project ⇐ repo.
+    if (!answers.operatorHandle) answers.operatorHandle = answers.owner;
+
+    const creation = await detectCreation(answers, skipGithub);
+    const project = resolveProjectDisplay(answers, skipGithub, projectsList);
     Logger.info(
-      JSON.stringify(
-        {
-          answers,
-          flags: state.flags,
-          manifest: buildMutationManifest(manifestCtx),
-        },
-        null,
-        2,
+      renderAnswerSummary(
+        answers,
+        creation,
+        project,
+        state.gitInitialized,
+        resolveRepoVisibility(state.flags),
       ),
     );
-    return { ok: false, exit: 0 };
+    const correct = await confirmYesNo('Is this correct?', state.interactive);
+    if (!correct) {
+      Logger.info('[bootstrap] Okay — let’s try again.');
+      // Re-prompt everything on the next pass (drop silent-accept).
+      silentAccept = [];
+      continue;
+    }
+
+    // In --dry-run we only collect/confirm info, so never ask to create.
+    if (!state.flags['dry-run'] && (creation.newRepo || creation.newProject)) {
+      const approved = await confirmYesNo(
+        'Create the new GitHub repo/project listed above?',
+        state.interactive,
+      );
+      if (!approved) {
+        Logger.error(
+          '[bootstrap] Creation declined — cannot continue without the repo/project. Exiting.',
+        );
+        return { ok: false, exit: 1 };
+      }
+    }
+    return { ok: true, payload: { answers, creation } };
   }
-  return { ok: true, payload: { answers } };
 }
 
 /**
- * Phase 3.5 — Manifest-first, phased approval (Story #3524). Renders the
- * FULL mutation manifest (every phase group, including the GitHub-admin
- * entries unless `--skip-github`) BEFORE collecting any approval, then
- * walks the four phase groups one at a time. Each group is independently
- * approvable — declining one never skips the others. `--assume-yes` pins
- * every prompt to "yes" (required for non-TTY runs); a non-TTY run without
- * `--assume-yes` records every group as declined (the HITL confirm refuses
- * to silent-apply without a TTY).
- *
- * Returns the approved phase-group `Set` plus the per-group decision log,
- * threaded into the execution phases as the single consent gate.
- *
- * Exported for tests.
- *
- * @param {{
- *   answers: object,
- *   assumeYes: boolean,
- *   flags: Record<string, string|boolean>,
- * }} state
- * @param {{ confirm?: Function }} [opts] — injectable confirm primitive.
- * @returns {Promise<PhaseAdvance>}
+ * --dry-run gate — print the resolved answers and the full mutation plan,
+ * then halt BEFORE any file write, GitHub change, or label creation. Runs
+ * after collect/confirm so the operator sees exactly what would happen.
  */
-export async function approvePhases(state, opts = {}) {
-  const manifestCtx = {
-    answers: state.answers,
-    skipGithub: Boolean(state.flags['skip-github']),
-    skipQuality: Boolean(state.flags['skip-quality']),
-  };
-  const { approved, decisions } = await runPhasedApproval({
-    ctx: manifestCtx,
-    log: (line) => Logger.info(line),
-    confirm: opts.confirm,
-    assume: state.assumeYes ? 'yes' : undefined,
-  });
-  return { ok: true, payload: { approvedGroups: approved, decisions } };
+/** Render the dry-run plan as a per-section layout (no mutations happen). */
+function renderDryRunPlan(state) {
+  const a = state.answers ?? {};
+  const c = state.creation ?? {};
+  const flagList = Object.entries(state.flags ?? {}).map(([k, v]) =>
+    v === true ? k : `${k}=${v}`,
+  );
+  return [
+    '\n=== Dry run — nothing will be changed ===',
+    'Values',
+    `  owner            ${a.owner ?? '(none)'}`,
+    `  operator handle  ${a.operatorHandle ?? '(none)'}`,
+    `  repo             ${a.repo ?? '(none)'}`,
+    `  base branch      ${a.baseBranch ?? '(none)'}`,
+    `  project number   ${a.projectNumber || '(skip)'}`,
+    '',
+    'Creation',
+    `  git init         ${state.gitInitialized ? 'no' : 'yes'}`,
+    `  new repo         ${c.newRepo ? `yes (${resolveRepoVisibility(state.flags)})` : 'no'}`,
+    `  new project      ${c.newProject ? 'yes' : 'no'}`,
+    '',
+    'Flags',
+    `  ${flagList.length ? flagList.join(', ') : '(none)'}`,
+  ].join('\n');
+}
+
+export function dryRunPlan(state) {
+  if (!state.flags['dry-run']) return { ok: true, payload: {} };
+  Logger.info(
+    '[bootstrap] --dry-run: no files, GitHub settings, or labels will be changed.',
+  );
+  Logger.info(renderDryRunPlan(state));
+  return { ok: false, exit: 0 };
 }
 
 /**
- * Phase 4 — Execute the project-side bootstrap and return the structured
- * report. Logs the start banner so operator-visible output is unchanged
- * from the pre-refactor inline pipeline. The approved phase-group gate
- * from {@link approvePhases} is threaded through so only consented
- * mutations land.
+ * Step 5 — Provision the missing pieces of a cold start, in dependency order:
  *
- * @param {{
- *   answers: object,
- *   projectRoot: string,
- *   agentRoot: string,
- *   flags: Record<string, string|boolean>,
- *   approvedGroups: Set<string>,
- * }} state
- * @returns {Promise<PhaseAdvance>}
+ *   1. Local git — `git init` + an initial commit when the folder is not a
+ *      repo yet (so the repo create below has something to push).
+ *   2. GitHub repo — `gh repo create --source --remote --push` when the repo
+ *      does not exist for the owner; otherwise wire the `origin` remote to the
+ *      existing repo when the local folder is not yet linked. Either way the
+ *      GitHub bootstrap can resolve the target from the local remote.
+ *   3. GitHub Project V2 — `gh project create` from the typed name when the
+ *      project answer is a name rather than an existing number; the assigned
+ *      number is written back onto `state.answers.projectNumber`.
+ *   4. Link — `gh project link` ties the repo to the project board so its
+ *      issues/PRs surface there (non-fatal; safe to re-run).
+ *
+ * Every action is idempotent and guarded by the detection done in
+ * `collectAndConfirm`, so a re-run on an already-provisioned project is a
+ * no-op. `--skip-github` suppresses the GitHub mutations but still runs the
+ * local git init. `--dry-run` never reaches this step (it halts earlier).
+ */
+export async function provisionResources(state) {
+  const skipGithub = Boolean(state.flags['skip-github']);
+
+  // 1. Local git — initialize + first commit when missing (idempotent).
+  const git = ensureGitInitialized(state);
+  if (!git.ok) {
+    Logger.error(`[bootstrap] git initialization failed: ${git.error}`);
+    return { ok: false, exit: 1 };
+  }
+  if (!git.initialized && !git.committed) {
+    Logger.info('[bootstrap] git already initialized — leaving as-is.');
+  }
+
+  const { newRepo, newProject } = state.creation;
+  if (skipGithub) {
+    if (newRepo || newProject) {
+      Logger.info(
+        '[bootstrap] --skip-github set; not creating the GitHub repo/project.',
+      );
+    }
+    return { ok: true, payload: {} };
+  }
+
+  // 2. GitHub repo — create + link + push when it does not exist yet;
+  //    otherwise ensure the local `origin` remote points at the existing repo
+  //    so the GitHub bootstrap can resolve the target (idempotent re-runs and
+  //    pre-created repos would otherwise leave the folder unlinked).
+  if (newRepo) {
+    try {
+      await createGithubRepo(state);
+    } catch (err) {
+      logGhError('repo create', err);
+      return { ok: false, exit: 1 };
+    }
+  } else {
+    await ensureGitRemote(state);
+  }
+
+  // 3. GitHub Project V2 — create from the typed name; capture its number so
+  //    the persist + GitHub bootstrap steps reuse it instead of duplicating.
+  if (newProject) {
+    try {
+      await createGithubProject(state);
+      // It now exists with a real number; downstream treats it as existing.
+      state.creation.newProject = false;
+    } catch (err) {
+      logGhError('project create', err);
+      return { ok: false, exit: 1 };
+    }
+  }
+
+  if (!newRepo && !newProject) {
+    Logger.info('[bootstrap] No new GitHub resources needed.');
+  }
+
+  // 4. Link the repo to the project board so issues/PRs surface on it
+  //    (idempotent + non-fatal; runs for both freshly created and existing
+  //    repo/project pairs).
+  await ensureProjectLinked(state);
+
+  return { ok: true, payload: {} };
+}
+
+/**
+ * Step 6a — Project-side bootstrap. With phased approval removed, all
+ * project-side phase groups are treated as approved.
  */
 export async function executeBootstrap(state) {
   Logger.info(
     `[bootstrap] Starting project bootstrap at ${state.projectRoot} (owner=${state.answers.owner} repo=${state.answers.repo} base=${state.answers.baseBranch})`,
   );
+  const approvedGroups = new Set(Object.values(PHASE_GROUPS));
   const report = await applyProjectBootstrap({
     projectRoot: state.projectRoot,
     agentRoot: state.agentRoot,
     answers: state.answers,
-    approvedGroups: state.approvedGroups,
+    approvedGroups,
     skipQuality: Boolean(state.flags['skip-quality']),
   });
-  return { ok: true, payload: { report } };
+  return { ok: true, payload: { report, approvedGroups } };
 }
 
 /**
- * Phase 5 — GitHub-side bootstrap. Honours `--skip-github`; on failure
- * the error is captured on the report rather than thrown so the summary
- * still prints. Always advances the pipeline so `printSummary` runs.
- *
- * The GitHub-side mutations are the irreversible `github-admin` phase
- * group; they run only when that group was approved in {@link approvePhases}
- * (in addition to the existing `--skip-github` opt-out). A declined
- * `github-admin` group records `{ skipped: true, reason }` on the report so
- * the summary reflects the consent decision.
- *
- * @param {{
- *   report: object,
- *   answers: object,
- *   flags: Record<string, string|boolean>,
- *   assumeYes: boolean,
- *   approvedGroups: Set<string>,
- * }} state
- * @returns {Promise<PhaseAdvance>}
+ * Step 6 (between project + GitHub) — Persist the chosen Projects V2 number
+ * into .agentrc.json's github block so it is the stored source of truth that
+ * resolveConfig (and the orchestrator) read back. Runs AFTER the project-side
+ * bootstrap has ensured .agentrc.json exists and BEFORE the GitHub bootstrap,
+ * so the provider reuses the existing project instead of creating a new one.
+ * Merges into an existing file (ensureAgentrc never overwrites one). Stored as
+ * an integer per the schema; a blank/new-project answer stores nothing.
  */
+export function persistProjectNumber(state) {
+  const pn = String(state.answers.projectNumber ?? '');
+  if (!/^\d+$/.test(pn)) {
+    return { ok: true, payload: {} };
+  }
+  const target = path.join(state.projectRoot, '.agentrc.json');
+  let config;
+  try {
+    config = JSON.parse(fs.readFileSync(target, 'utf8'));
+  } catch (err) {
+    Logger.error(
+      `[bootstrap] Could not read ${target} to store projectNumber: ${err.message}`,
+    );
+    return { ok: true, payload: {} };
+  }
+  config.github = config.github ?? {};
+  if (config.github.projectNumber !== Number(pn)) {
+    config.github.projectNumber = Number(pn);
+    fs.writeFileSync(target, `${JSON.stringify(config, null, 2)}\n`, 'utf8');
+    Logger.info(
+      `[bootstrap] Stored github.projectNumber=${pn} in .agentrc.json`,
+    );
+  }
+  return { ok: true, payload: {} };
+}
+
+/** Step 6b — GitHub-side bootstrap. Honours `--skip-github`. */
 export async function executeGithubBootstrap(state) {
-  if (state.flags['skip-github']) return { ok: true, payload: {} };
-  if (
-    state.approvedGroups &&
-    !state.approvedGroups.has(PHASE_GROUPS.GITHUB_ADMIN)
-  ) {
-    state.report.github = {
-      skipped: true,
-      reason: 'phase-group-declined',
-    };
+  if (state.flags['skip-github']) {
+    Logger.info('[bootstrap] --skip-github set; skipping GitHub bootstrap.');
     return { ok: true, payload: {} };
   }
   try {
     state.report.github = await runGithubBootstrap(state.answers, {
       assumeYes: state.assumeYes,
+      reapConflictingWorkflows: Boolean(
+        state.flags['reap-conflicting-workflows'],
+      ),
     });
   } catch (err) {
-    Logger.error(`[bootstrap] GitHub bootstrap failed: ${err.message}`);
+    // GhExecError carries the real gh stderr/stdout/exit code — surface it so
+    // a generic "gh exited with code 1" is actually diagnosable.
+    logGhError('GitHub bootstrap', err);
     state.report.github = { error: err.message };
   }
   return { ok: true, payload: {} };
 }
 
-/**
- * Compute the phase groups whose mutations actually landed, for the install
- * ledger. Project-side groups (`ide-wiring`, `repo-config`, `quality-gates`)
- * count as applied when they were approved. The remote `github-admin` group
- * counts as applied only when it was approved AND the GitHub bootstrap
- * succeeded — a skipped or errored GitHub run never records irreversible
- * remote mutations in the ledger that a future `mandrel uninstall` would try
- * to roll back.
- *
- * Exported for tests.
- *
- * @param {Set<string>} approvedGroups
- * @param {object} report
- * @returns {Set<string>}
- */
-export function resolveAppliedGroups(approvedGroups, report) {
-  const applied = new Set();
-  for (const group of approvedGroups ?? []) {
-    if (group === PHASE_GROUPS.GITHUB_ADMIN) {
-      const gh = report?.github;
-      if (gh && !gh.error && !gh.skipped && githubSubMutationsSucceeded(gh))
-        applied.add(group);
-      continue;
-    }
-    applied.add(group);
-  }
-  return applied;
-}
-
-/**
- * Return true only when every github-admin sub-mutation that ran either
- * succeeded or was intentionally skipped (opt-out / no-config). A `status`
- * of `'failed'` on any sub-result means at least one remote mutation did not
- * land, so the ledger must NOT record the `github-admin` group as applied.
- *
- * @param {object} gh — the `report.github` object after a non-error, non-skip run.
- * @returns {boolean}
- */
-function githubSubMutationsSucceeded(gh) {
-  if (gh.branchProtection?.status === 'failed') return false;
-  if (gh.mergeMethods?.status === 'failed') return false;
-  return true;
-}
-
-/**
- * Phase 6 — Record the install ledger (Story #3524). Writes
- * `.agents/.install-manifest.json` enumerating exactly the approved-and-
- * applied mutation-manifest entries so a future `mandrel uninstall` can
- * consume it. The ledger is gitignored. When no group landed (e.g. a
- * non-TTY run that declined everything), no ledger is written.
- *
- * Exported for tests.
- *
- * @param {{
- *   answers: object,
- *   projectRoot: string,
- *   flags: Record<string, string|boolean>,
- *   approvedGroups: Set<string>,
- *   report: object,
- * }} state
- * @returns {PhaseAdvance}
- */
+/** Step 6c — Record the install ledger for a future uninstall. */
 export function recordLedger(state) {
   const appliedGroups = resolveAppliedGroups(
     state.approvedGroups,
@@ -832,17 +1079,7 @@ export function recordLedger(state) {
   return { ok: true, payload: {} };
 }
 
-/**
- * Pipeline driver — chains a sequence of phase helpers, threading the
- * accumulated state through each call. The first helper to return
- * `{ ok: false }` short-circuits and the driver returns `result.exit`.
- *
- * Exported for tests so the contract is asserted without spawning a
- * child process.
- *
- * @param {Array<(state: object) => Promise<object>|object>} phases
- * @returns {Promise<{ ok: boolean, exit?: number, state: object }>}
- */
+/** Pipeline driver — threads accumulated state through each phase. */
 export async function runPipeline(phases) {
   let state = {};
   for (const phase of phases) {
@@ -853,19 +1090,20 @@ export async function runPipeline(phases) {
   return { ok: true, state };
 }
 
-export async function main(argv = process.argv.slice(2), opts = {}) {
+export async function main(argv = process.argv.slice(2)) {
   const result = await runPipeline([
     () => parseAndValidate(argv),
-    (s) => runPreflightPhase(s, { run: opts.preflightRun }),
     (s) => prepareContext(s),
-    (s) => collectAndValidateAnswers(s),
-    (s) => approvePhases(s, { confirm: opts.confirm }),
+    (s) => runPreflightPhase(s),
+    (s) => collectAndConfirm(s),
+    (s) => dryRunPlan(s),
+    (s) => provisionResources(s),
     (s) => executeBootstrap(s),
+    (s) => persistProjectNumber(s),
     (s) => executeGithubBootstrap(s),
     (s) => recordLedger(s),
   ]);
   if (!result.ok) return result.exit;
-  printSummary(result.state.report);
   Logger.info('\n[bootstrap] Done.');
   return 0;
 }
