+++ .agents/scripts/lib/bootstrap/project-bootstrap.js
@@ -16,7 +16,6 @@ import { spawnSync as defaultSpawnSync } from 'node:child_process';
 import fs from 'node:fs';
 import os from 'node:os';
 import path from 'node:path';
-import { buildProfileAgentrcBody } from '../config/sync-agentrc.js';
 import { PHASE_GROUPS, previewMutationManifest } from './manifest.js';
 import { applyQualityBootstrap } from './quality-bootstrap.js';
 
@@ -245,41 +244,23 @@ export function ensureDependenciesInstalled(ctx) {
 /**
  * Step 2.5a — Seed `.agentrc.json` when missing.
  *
- * Two seeding sources, selected by whether the operator picked a named
- * config profile (Story #3527):
+ * Seeds from the bundled `starter-agentrc.json` reference, applying the
+ * operator-identity placeholder substitution (`[OWNER]` / `[REPO]` /
+ * `[USERNAME]`) and the `baseBranch` override. (Story #3690 removed the
+ * named config-profile seeding path — the starter reference is the single
+ * seed source.)
  *
- *   - **Profile selected** (`ctx.answers.profile` is a known profile name):
- *     seed from that profile's delta via
- *     {@link buildProfileAgentrcBody}. The minimal `solo-local` profile
- *     yields a correspondingly minimal `.agentrc.json` scoped to that
- *     posture — the repo-config phase seeds from the chosen profile delta
- *     rather than the full reference config.
- *   - **No profile** (`ctx.answers.profile` blank/absent): seed from the
- *     bundled `starter-agentrc.json` reference (the historical default).
- *
- * Both sources apply the same operator-identity placeholder substitution
- * (`[OWNER]` / `[REPO]` / `[USERNAME]`) and `baseBranch` override.
- *
- * An existing `.agentrc.json` is never overwritten — operator wins. A
- * profile name that fails to resolve/validate propagates as a throw so the
- * fatal `validation` phase never runs against a half-written file.
+ * An existing `.agentrc.json` is never overwritten — operator wins.
  *
  * @param {object} ctx
  * @param {typeof fs} [ctx.fsImpl]
- * @returns {{ action: string, path: string, source?: string,
- *   profile?: string }}
+ * @returns {{ action: string, path: string, source?: string }}
  */
 export function ensureAgentrc(ctx) {
   const { fsImpl = fs } = ctx;
   const target = path.join(ctx.projectRoot, '.agentrc.json');
   if (fsImpl.existsSync(target)) {
     return { action: 'already-present', path: target };
-  }
-  const profile = ctx.answers.profile;
-  if (typeof profile === 'string' && profile.length > 0) {
-    const body = buildProfileAgentrcBody({ profile, answers: ctx.answers });
-    fsImpl.writeFileSync(target, body, 'utf8');
-    return { action: 'seeded', path: target, source: 'profile', profile };
   }
   const starter = path.join(
     ctx.agentRoot ?? path.join(ctx.projectRoot, '.agents'),
