+++ lib/cli/explain.js
@@ -3,17 +3,16 @@
  * `mandrel explain` subcommand.
  *
  * Prints, for every resolved config key, its effective value, the source
- * layer the value came from (default / profile / agentrc), and a one-line
- * meaning. Secret-shaped keys report `<redacted>` in place of a value — only
- * their source is shown.
+ * layer the value came from (default / agentrc), and a one-line meaning.
+ * Secret-shaped keys report `<redacted>` in place of a value — only their
+ * source is shown.
  *
  * Usage:
  *   mandrel explain                 # explain the project's resolved config
- *   mandrel explain --profile <p>   # attribute omitted keys to a named profile
  *   mandrel explain --json          # emit the raw report as JSON
  *
  * Output goes to process.stdout so it is capturable by tests. Exit code 0 on
- * success, 1 on an unknown profile or resolution failure.
+ * success, 1 on a resolution failure.
  *
  * Injectable seams (used by tests):
  *   - `explain` — replaces the default `explainConfig` implementation
@@ -26,7 +25,6 @@ import { explainConfig } from '../../.agents/scripts/lib/config/explain.js';
 
 const SOURCE_BADGES = Object.freeze({
   agentrc: '[agentrc]',
-  profile: '[profile]',
   default: '[default]',
 });
 
@@ -46,23 +44,14 @@ function formatRow(entry) {
  * Parse the subcommand argv into options.
  *
  * @param {string[]} argv
- * @returns {{ profile: string|null, json: boolean }}
+ * @returns {{ json: boolean }}
  */
 export function parseArgs(argv) {
-  let profile = null;
   let json = false;
-  for (let i = 0; i < argv.length; i += 1) {
-    const arg = argv[i];
-    if (arg === '--json') {
-      json = true;
-    } else if (arg === '--profile') {
-      profile = argv[i + 1] ?? null;
-      i += 1;
-    } else if (arg.startsWith('--profile=')) {
-      profile = arg.slice('--profile='.length);
-    }
+  for (const arg of argv) {
+    if (arg === '--json') json = true;
   }
-  return { profile, json };
+  return { json };
 }
 
 /**
@@ -86,11 +75,11 @@ export function runExplain(
     exit = (code) => process.exit(code),
   } = {},
 ) {
-  const { profile, json } = parseArgs(argv);
+  const { json } = parseArgs(argv);
 
   let report;
   try {
-    report = explain({ profile });
+    report = explain({});
   } catch (err) {
     errOut(`mandrel explain: ${err.message}\n`);
     exit(1);
