+++ tests/scripts/bootstrap-question-pickers.test.js
@@ -1,22 +1,21 @@
 /**
- * bootstrap-question-pickers.test — Story #3377
+ * bootstrap-question-pickers.test — Story #3377, repointed at the collapsed
+ * orchestrator by Story #3690.
  *
  * Pins the picker wiring on the bootstrap question list:
  *
- *   1. `buildQuestions` attaches a `picker: { list }` to the `repo` question
- *      (backed by `listRepos`, mapping each `owner/name` slug to its bare
- *      repo name) and to the `projectNumber` question (backed by
- *      `listProjects`, returning project titles). The owner is resolved up
- *      front via the same flag → env → inferred-default precedence the
- *      `owner` question walks, so the pickers scope their `gh` queries to the
- *      right owner.
+ *   1. `buildQuestions` attaches a `picker: { list }` to the `repo` and
+ *      `projectNumber` questions. The pre-fetched `lists` (shared with the
+ *      summary display) are the fast path; when they are empty the picker
+ *      falls back to a live fetch keyed off the owner resolved so far —
+ *      and degrades to an empty list (skip) when no owner resolves.
  *   2. Precedence holds: when `--owner`/`--repo` or `GH_OWNER`/`GH_REPO` are
  *      supplied, or `--assume-yes` is set, the picker never runs because the
  *      earlier resolvers (`resolveFromFlag` / `resolveFromEnv` /
  *      `resolveAssumeYes`) win first — the supplied values are used verbatim.
  *
- * The `gh-list` providers are injected as stubs so no real `gh` child is
- * spawned (unit tier — all I/O is mocked).
+ * The pre-fetched lists are plain in-memory fixtures so no real `gh` child
+ * is spawned (unit tier — all I/O is mocked).
  */
 
 import assert from 'node:assert/strict';
@@ -38,27 +37,15 @@ const DEFAULTS = Object.freeze({
   operatorHandle: 'octo',
 });
 
+const PROJECTS = Object.freeze([
+  { label: 'Roadmap (#7)', value: '7' },
+  { label: 'Bugs (#9)', value: '9' },
+]);
+
 function findQuestion(questions, key) {
   const q = questions.find((entry) => entry.key === key);
   assert.ok(q, `expected a "${key}" question`);
   return q;
-}
-
-function stubProviders({ repos = [], projects = [] } = {}) {
-  const calls = { listRepos: [], listProjects: [] };
-  return {
-    calls,
-    providers: {
-      listRepos: (opts) => {
-        calls.listRepos.push(opts);
-        return repos;
-      },
-      listProjects: (opts) => {
-        calls.listProjects.push(opts);
-        return projects;
-      },
-    },
-  };
 }
 
 // ---------------------------------------------------------------------------
@@ -98,62 +85,47 @@ describe('resolveOwnerForPicker', () => {
 // ---------------------------------------------------------------------------
 
 describe('buildQuestions picker attachment', () => {
-  it('attaches a repo picker that lists bare repo names for the owner', () => {
-    const { calls, providers } = stubProviders({
-      repos: ['acme/api', 'acme/web'],
-    });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: {},
-      env: {},
-      providers,
-    });
+  it('attaches a repo picker that serves the pre-fetched bare repo names', () => {
+    const questions = buildQuestions(
+      DEFAULTS,
+      {},
+      {},
+      { reposList: ['api', 'web'], projectsList: [] },
+    );
     const repo = findQuestion(questions, 'repo');
     assert.equal(typeof repo.picker?.list, 'function');
-    assert.deepEqual(repo.picker.list(), ['api', 'web']);
-    // owner was scoped from the inferred default
-    assert.deepEqual(calls.listRepos[0], { owner: 'acme' });
+    assert.deepEqual(repo.picker.list({}), ['api', 'web']);
   });
 
-  it('leaves a slug without a slash untouched in the repo picker', () => {
-    const { providers } = stubProviders({ repos: ['solo'] });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: {},
-      env: {},
-      providers,
-    });
-    assert.deepEqual(findQuestion(questions, 'repo').picker.list(), ['solo']);
-  });
-
-  it('attaches a project picker that lists project titles for the owner', () => {
-    const { calls, providers } = stubProviders({
-      projects: ['Roadmap', 'Bugs'],
-    });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: {},
-      env: {},
-      providers,
-    });
+  it('attaches a project picker that serves the pre-fetched project choices', () => {
+    const questions = buildQuestions(
+      DEFAULTS,
+      {},
+      {},
+      { reposList: [], projectsList: PROJECTS },
+    );
     const project = findQuestion(questions, 'projectNumber');
     assert.equal(typeof project.picker?.list, 'function');
-    assert.deepEqual(project.picker.list(), ['Roadmap', 'Bugs']);
-    assert.deepEqual(calls.listProjects[0], { owner: 'acme' });
-  });
-
-  it('scopes the repo picker to the --owner flag when supplied', () => {
-    const { calls, providers } = stubProviders({ repos: ['other/x'] });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: { owner: 'other' },
-      env: {},
-      providers,
-    });
-    findQuestion(questions, 'repo').picker.list();
-    assert.deepEqual(calls.listRepos[0], { owner: 'other' });
+    assert.deepEqual(project.picker.list({}), PROJECTS);
   });
 
   it('does not attach a picker to the owner / baseBranch questions', () => {
-    const questions = buildQuestions(DEFAULTS, { flags: {}, env: {} });
+    const questions = buildQuestions(DEFAULTS, {}, {}, {});
     assert.equal(findQuestion(questions, 'owner').picker, undefined);
     assert.equal(findQuestion(questions, 'baseBranch').picker, undefined);
+  });
+
+  it('repo picker degrades to an empty list when no owner resolves and no list was pre-fetched', () => {
+    const questions = buildQuestions(
+      { ...DEFAULTS, owner: null },
+      {},
+      {},
+      { reposList: [], projectsList: [] },
+    );
+    const repo = findQuestion(questions, 'repo');
+    // No owner from defaults/flags/env and none answered yet → [] without
+    // any live gh fetch.
+    assert.deepEqual(repo.picker.list({}), []);
   });
 });
 
@@ -169,14 +141,29 @@ describe('picker precedence in collectAnswers', () => {
     output: { write: () => {} },
   });
 
-  it('uses the --repo flag verbatim and never calls the repo picker', async () => {
-    const { calls, providers } = stubProviders({ repos: ['acme/should-not'] });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: { owner: 'acme', repo: 'flag-repo' },
-      env: {},
-      providers,
+  function trackedQuestions(flags, env) {
+    let pickerCalls = 0;
+    const questions = buildQuestions(DEFAULTS, flags, env, {
+      reposList: ['should-not'],
+      projectsList: PROJECTS,
     });
-    const repo = questions.find((q) => q.key === 'repo');
+    for (const q of questions) {
+      if (!q.picker) continue;
+      const orig = q.picker.list;
+      q.picker.list = (answers) => {
+        pickerCalls += 1;
+        return orig(answers);
+      };
+    }
+    return { questions, count: () => pickerCalls };
+  }
+
+  it('uses the --repo flag verbatim and never calls the repo picker', async () => {
+    const { questions, count } = trackedQuestions(
+      { owner: 'acme', repo: 'flag-repo' },
+      {},
+    );
+    const repo = findQuestion(questions, 'repo');
     const { answers } = await collectAnswers({
       questions: [repo],
       flags: { repo: 'flag-repo' },
@@ -185,14 +172,13 @@ describe('picker precedence in collectAnswers', () => {
       ...noPromptIo,
     });
     assert.equal(answers.repo, 'flag-repo');
-    assert.equal(calls.listRepos.length, 0, 'picker.list must not be invoked');
+    assert.equal(count(), 0, 'picker.list must not be invoked');
   });
 
   it('uses GH_REPO env verbatim and never calls the repo picker', async () => {
-    const { calls, providers } = stubProviders({ repos: ['acme/should-not'] });
     const env = { GH_OWNER: 'acme', GH_REPO: 'env-repo' };
-    const questions = buildQuestions(DEFAULTS, { flags: {}, env, providers });
-    const repo = questions.find((q) => q.key === 'repo');
+    const { questions, count } = trackedQuestions({}, env);
+    const repo = findQuestion(questions, 'repo');
     const realEnv = process.env;
     process.env = { ...realEnv, ...env };
     try {
@@ -204,22 +190,14 @@ describe('picker precedence in collectAnswers', () => {
         ...noPromptIo,
       });
       assert.equal(answers.repo, 'env-repo');
-      assert.equal(calls.listRepos.length, 0);
+      assert.equal(count(), 0);
     } finally {
       process.env = realEnv;
     }
   });
 
   it('with --assume-yes accepts defaults and never calls the pickers', async () => {
-    const { calls, providers } = stubProviders({
-      repos: ['acme/should-not'],
-      projects: ['Should Not'],
-    });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: {},
-      env: {},
-      providers,
-    });
+    const { questions, count } = trackedQuestions({}, {});
     const { answers } = await collectAnswers({
       questions,
       flags: {},
@@ -228,17 +206,16 @@ describe('picker precedence in collectAnswers', () => {
     });
     assert.equal(answers.owner, 'acme');
     assert.equal(answers.repo, 'inferred-repo');
-    assert.equal(calls.listRepos.length, 0);
-    assert.equal(calls.listProjects.length, 0);
+    assert.equal(count(), 0);
   });
 
-  it('renders the repo picker and returns the selected bare name when interactive', async () => {
-    const { providers } = stubProviders({ repos: ['acme/api', 'acme/web'] });
-    const questions = buildQuestions(DEFAULTS, {
-      flags: {},
-      env: {},
-      providers,
-    });
+  it('renders the repo picker and returns the selected name when interactive', async () => {
+    const questions = buildQuestions(
+      DEFAULTS,
+      {},
+      {},
+      { reposList: ['api', 'web'], projectsList: [] },
+    );
     const repo = findQuestion(questions, 'repo');
     const lines = [];
     const ctx = {
@@ -259,13 +236,12 @@ describe('picker precedence in collectAnswers', () => {
   });
 
   it('skips the repo picker when no owner can be resolved (empty list)', async () => {
-    const { calls, providers } = stubProviders({ repos: [] });
-    const defaults = { ...DEFAULTS, owner: null };
-    const questions = buildQuestions(defaults, {
-      flags: {},
-      env: {},
-      providers,
-    });
+    const questions = buildQuestions(
+      { ...DEFAULTS, owner: null },
+      {},
+      {},
+      { reposList: [], projectsList: [] },
+    );
     const repo = findQuestion(questions, 'repo');
     const ctx = {
       q: repo,
@@ -278,8 +254,5 @@ describe('picker precedence in collectAnswers', () => {
       output: { write: () => {} },
     };
     assert.deepEqual(await resolveFromPicker(ctx), { kind: 'skip' });
-    // owner is null, listRepos stub returns [] → no menu rendered.
-    assert.equal(calls.listRepos.length, 1);
-    assert.deepEqual(calls.listRepos[0], { owner: null });
   });
 });
